Fix - type date BigInt

+++ src/Firebird/Schema/Grammars/FirebirdGrammar.php
@@ -287,7 +287,7 @@ class FirebirdGrammar extends Grammar {
    */
   protected function typeBigInteger(Fluent $column)
   {
-    return 'INTEGER';
+    return 'BIGINT';
   }
 
   /**
